Use a ProximityDetector
