+++ src/main/scala/org/clulab/wm/eidos/groundings/grounders/srl/SRLCompositionalGrounder.scala
@@ -31,6 +31,7 @@ case class GroundedSpan(tokenInterval: Interval, grounding: OntologyGrounding, i
 class SRLCompositionalGrounder(name: String, domainOntology: DomainOntology, w2v: EidosWordToVec, canonicalizer: Canonicalizer, tokenizer: EidosTokenizer)
     extends EidosOntologyGrounder(name, domainOntology, w2v, canonicalizer) {
 
+  val proximityDetector = new RightLeftProximityDetector()
   val emptyPredicateTuple: PredicateTuple = PredicateTuple(emptyOntologyGrounding, emptyOntologyGrounding, emptyOntologyGrounding, emptyOntologyGrounding)
 
   lazy val proc: CluProcessor = {
@@ -176,75 +177,11 @@ class SRLCompositionalGrounder(name: String, domainOntology: DomainOntology, w2v
       }
     }
 
-    def indexAndCountToCount(position: Int, leftIndexAndCountOpt: Option[(Int, Int)], rightIndexAndCountOpt: Option[(Int, Int)]): Int = {
-      (leftIndexAndCountOpt, rightIndexAndCountOpt) match {
-        case (None, None) => 0
-        case (None, Some((_, rightCount))) => rightCount
-        case (Some((_, leftCount)), None) => leftCount
-        case (Some((leftIndex, leftCount)), Some((rightIndex, rightCount))) =>
-          // Whichever is closest to the position and when tied, the value that
-          // is higher than position because we favor right, even though it might
-          // be the higher numbered arg or pred.
-          val leftDist = math.abs(position - leftIndex)
-          val rightDist = math.abs(position - rightIndex)
-          if (leftDist < rightDist) leftCount
-          else if (leftDist == rightDist) {
-            // These should not ever be equal because then there would be an
-            // arg and pred at the same position.
-            if (leftIndex > rightIndex) leftIndex
-            else rightIndex
-          }
-          else rightCount
-      }
-    }
-
     def getBest(indexes: Seq[Int], ontologyGroundingOpts: Seq[Option[OntologyGrounding]]): Option[OntologyGrounding] = {
       if (indexes.isEmpty) None
       else {
         val ontologyGroundings = indexes.map(ontologyGroundingOpts(_).get)
         Some(ontologyGroundings.maxBy(_.individualGroundings.head.score))
-      }
-    }
-
-    def findWhereAndWhatOptAround[T](values: Seq[T], position: Int)(f: T => Boolean): Option[(Int, T)] = {
-      val whereAndWhatOptBefore = Collection.findWhereAndWhatOptBefore(values, position)(f)
-      val whereAndWhatOptAfter = Collection.findWhereAndWhatOptAfter(values, position)(f)
-
-      (whereAndWhatOptBefore, whereAndWhatOptAfter) match {
-        case (None, None) => None
-        case (None, some) => some
-        case (some, None) => some
-        case (Some(before), Some(after)) =>
-          // Prefer the one after in case of ties.
-          if (math.abs(after._1 - position) <= math.abs(before._1 - position)) Some(after)
-          else Some(before)
-      }
-    }
-
-    def getAroundCount(index: Int, intArgs: Seq[Int], intPreds: Seq[Int]): Int = {
-      val argWhereAndWhatOpt: Option[(Int, Int)] = findWhereAndWhatOptAround(intArgs, index)(_ != 0)
-      val predWhereAndWhatOpt: Option[(Int, Int)] = findWhereAndWhatOptAround(intPreds, index)(_ != 0)
-
-      if (argWhereAndWhatOpt.isDefined || predWhereAndWhatOpt.isDefined)
-        indexAndCountToCount(index, argWhereAndWhatOpt, predWhereAndWhatOpt)
-      else
-        0
-    }
-
-    def getRightThenLeftCount(index: Int, intArgs: Seq[Int], intPreds: Seq[Int]): Int = {
-      // Prefer looking to the right, after.
-      val argWhereAndWhatOpt: Option[(Int, Int)] = Collection.findWhereAndWhatOptAfter(intArgs, index)(_ != 0)
-      val predWhereAndWhatOpt: Option[(Int, Int)] = Collection.findWhereAndWhatOptAfter(intPreds, index)(_ != 0)
-      if (argWhereAndWhatOpt.isDefined || predWhereAndWhatOpt.isDefined)
-        indexAndCountToCount(index, argWhereAndWhatOpt, predWhereAndWhatOpt)
-      else {
-        // Then look left if necessary, before.
-        val argWhereAndWhatOpt = Collection.findWhereAndWhatOptBefore(intArgs, index)(_ != 0)
-        val predWhereAndWhatOpt = Collection.findWhereAndWhatOptBefore(intPreds, index)(_ != 0)
-        if (argWhereAndWhatOpt.isDefined || predWhereAndWhatOpt.isDefined) {
-          indexAndCountToCount(index, argWhereAndWhatOpt, predWhereAndWhatOpt)
-        }
-        else 0
       }
     }
 
@@ -275,8 +212,7 @@ class SRLCompositionalGrounder(name: String, domainOntology: DomainOntology, w2v
         // -n, property that belongs to a pred
         val intProps: Seq[Int] = indices.map { index =>
           if (!isProps(index)) 0
-          else getRightThenLeftCount(index, intArgs, intPreds)
-          // else getAroundCount(index, intArgs, intPreds)
+          else proximityDetector.detect(index, intArgs, intPreds)
         }
         // There is a property and it maps to no arg or pred.
         val zeroPropertyIndexes = indices.filter { index => isProps(index) && intProps(index) == 0 }
@@ -519,15 +455,17 @@ class SRLCompositionalGrounder(name: String, domainOntology: DomainOntology, w2v
       SRLCompositionalGrounder.triggerPartsOfSpeech.exists(pos.startsWith)
     }
 
-    mention.attachments.flatMap { a =>
-      a match {
+    val attachmentStrings = mention.attachments.flatMap { attachment =>
+      attachment match {
         case _: Property => Seq.empty
-        case t: TriggeredAttachment =>
-          Seq(t).filter(qualifies).map(_.trigger) ++ t.quantifiers.getOrElse(Seq())
-        case c: ContextAttachment => Seq(c.text)
+        case triggeredAttachment: TriggeredAttachment =>
+          Seq(triggeredAttachment).filter(qualifies).map(_.trigger) ++ triggeredAttachment.quantifiers.getOrElse(Seq.empty)
+        case contextAttachment: ContextAttachment => Seq(contextAttachment.text)
         case _ => throw new RuntimeException("Unexpected attachment")
       }
     }
+    // println(s"attachmentStrings: $attachmentStrings")
+    attachmentStrings
   }
 
   // In SRL, currently the arguments that point to a preposition stay there, rather than continuing
